+++ CHANGELOG.md
@@ -7,10 +7,11 @@ This is a major release with **important breaking changes**, please see our [mig
 - Removed Expo support from this repository. You should now use our new dedicated [Batch-Expo-Plugin](https://github.com/BatchLabs/Batch-Expo-Plugin). 
 
 **Core**
+- Removed support for the old React Native architecture (Native Module). The plugin is now a pure Turbo Module and requires the new architecture enabled.
 - Batch no longer requires a custom React Native CLI configuration. If `react-native.config.js` only exists for Batch, delete it or remove the `@batch.com/react-native-plugin` entry.
 
 **Android**
-- The plugin is no longer auto-initialized. You should now call `RNBatchModuleImpl.initialize(application)` from `MainApplication.onCreate()` to complete setup.
+- The plugin is no longer auto-initialized. You should now call `RNBatchModule.initialize(application)` from `MainApplication.onCreate()` to complete setup.
 - The initial state of the "Do Not Disturb" (DnD) feature is no longer read from the Android resources. You should now add a meta-data tag `batch_do_not_disturb_initial_state` to the <application> section of your `AndroidManifest`.